Include some new cache control settings
After playing a little bit with trying to get the data smaller in some
fashion or otherwise speeding up the operation, the final thing to do
with this now is control whether the cache happens and whether or not it
encrypts, and call it good.

To that end, we now have two settings; one to indicate if caching should
happen at all, and another to control if it should be encrypted or not.

For our purposes in testing, these default to cacheing but not
encrypting for now, which is fast and gets us going quicker.

At some point in the future, we can use a proper encryption library,
which should speed this up considerably.