Use the new settings in the cache code
Using the new settings, modify the load and save routines in the cache
to optionally do nothing, or load and possibly decrypt before we process
the incoming data.

All combinations of this have been tested at work, though for full
completeness we should watch the settings object and automatically
invalidate the cache data when the setting changes, since currently
that must be done manually.