Factor video browsing code into the core code
The code for prompting for a video is going to be used in at least two
places, so it has been hoisted in this commit and refactored to be a
new common selection routine similar to the previous YouTubeRequest
class.

In this flow, the only thing that need be implemented to support the
browse is the method that gets told when a video is selected (which by
default prompts for timecode) and when a video toc entry is chosen.