Make placeholders dynamic in the new abstraction
This change modifies the previous refactored code so that subclasses
have control over what the placeholders say in the quick panel when
prompting for various bits of information.

The default is to defer to the underlying utility code that does the
prompting (which has had its defaults changed slightly), but these can
be modified as desired.

The command to get a link takes advantage of this to make sure that all
through the process of choosing an item, you know what you're picking.