Fix a bug in yesterday's last commit
The last commit from yesterday, which added the whole video data as a 
setting to a YouTube video editor window, had an issue in it that needed
to be resolved.

In particular the returned data when a query is made is only wrapped as
a dotty in the outer later (the main object) and not on any of the inner
keys, but when the cache data is loaded back from the JSON, the decoder
converts all dictionaries to a dotty.

This blocks the ability to add this data as a setting or pass it as an
argument because it contains a non-primitive value.

To resolve this issue, we implemented a simple "undotty" that can
recursively unwrap the value of a dotty and return back the base
dictionary.

Since the Dotty library keeps a copy of the original dict it was added
to, this has a very low resource cost.