New command for opening a video in YouTube Studio
A new command has been created and added to the context menu of the
body pane in a YouTube editor window which creates a link to edit the
video in YouTube studio and then opens the link in a tab.

This will be a handy way to get at the core UI of Studio as needed, and
also a stopgap so that we can do actual editing easier before we get our
own functionality completed.