New internal code for displaying utility output
The last change we're working on for this evening involves the idea of
having "Utility Functions" which can generate logs of videos that meet
certain requirements, such as not having a Table of Contents, having
text in their description, and so on.

To support that, we have added in the ability to get at a report view
and add text to it. 

This is smart enough to be able to use either a regular view or a panel,
which is controlled by a setting (which defaults to the panel).