New utility to gather the TOC for a video
This includes a new utility function that, given a video object, will
return back a list of tuples that provide the table of contents entries
for the video (the time code and the text).

This is currently used in the place the code was factored from, but
will momentarily be used in a new utility command.