Include Missing TOC support files
The first utility function that we're going to add is one which can
provide a list of videos for which there is no Table of Contents
defined in the description.

To that end, we have a syntax that applies some simple styles to make
viewing either, syntax specific settings appropriate for this, as well
as a menu to be displayed in the output (be it a panel or a view).