Update the syntax for the missing TOC pane
This is a syntax tweak that updates how we syntax highlight the results
of our missing TOC utility text.

In particular this makes the title a bit nicer by allowing for a
separator, and starts the notion that items that we output should
have their video ID prepended to them in a specific format that can
later be post-processed away.