Include a new meta command
This includes a new command that can grab information out of the buffer
based on the newly applied scope from the syntax, and replace it with
some inline phantoms that provide controls based on that information.

For now the only part of this that does anything is to remove away the
information from the record.