Execute the new meta command as a post-processor
To tie all of the most recent changes together, after generating the
report this executes the new meta command to cause the video records to
be replaced with video controls.

As of yet the controls are not specified, so what this mostly represents
is getting the output back to what it looked like before we started
doing anything this evening.