Update the new meta command to add phantoms
Modify our new meta command so that instead of just removing the new
video information, it actually inserts some HTML content in place.

This requires a slight tweak to the original simplified command in that
in order to make the changes we want, we need to iterate over the list
of found regions in a forward direction instead of a backwards one,
since we need the regions that we use to apply the phantoms to be the
final location of each text item.