Change how we apply our phantoms
After finishing up an epic 3.5 hour live stream, more or less the only
thing that came out of it was that the way we were using phantoms, while
it seems to be correct, seems to result in phantoms not being visually
removed from the buffer even though they appear in all regards to be
gone when querying the API.

(For future purposes: this was as related to the plugin reloading
causing existing phantoms to remain behind even though the code seemed
to be clearing them if we implemented __del__; they would stay behind
until deleting them by their key text).

This seems to be related to PhantomSet and something that it's doing (or
not doing) behind the scenes, or it is perhaps tweaking something in the
API since it works with phantoms based on their ID.

Here we know that we're only applying phantoms in a single shot and we
will never manipulate the list, so rather than using PhantomSet and id
values, we're just manually adding regions and redacting them away.