Update the syntax to support the new layout
This is only a test layout since we're most likely going to be using a
more simple display and a hover popup, but for the time being this adds
in some extra visual juice (and it was nice to actually implement
anything after things went so wrong with phantoms).

Here the syntax now supports and scopes the text on lines based on the
new report layout. The scopes and the match are rather crude, but serve
for the demonstration purposes.

For reports like this, we're also turning off spell checking by default,
since in theory these videos have already gone through the normal
editing process and would have known typos fixed. It's easy enough to
turn on as needed.