Re-introduce our sorting utility helper function
In order to make some bits of code a bit cleaner, this re-implements a
new and enhanced version of a previous helper. This helper presumes
that you're giving it a sorted list of videos, and will return a sorted
version of the list.

The sort needs to be provided only the key to sort on, and can
optionally include a type function for coercion during the sort as well
as an optional flag to reverse the sort.

In the future this will also get filtering support as well, which makes
it slightly better that this simple one line of code is being put into
a utility just to make the call point slightly better looking.