Make the report output focus the view on update
In the code we wrote to generate output to either a panel or a view
(depending on the setting) we implemented code that makes the panel
appear when the output is added, but what we didn't do was allow for
the view to be focused if the setting says to put output there.

During testing this was incredibly annoying, so lets fix that before we
go any further.