Implement a lookup for video hover information
This implements the logic necessary to get at the information on what
video is being hovered over so that it can be displayed in a popup.

This works by determining the region that spans the title that is under
the cursor and then checking to see which of the regions that contain
that scope this is, which gives us an index that says which title it is
starting from the top of the buffer.

From here we can use that index to get the index of the appropriate
video id from the settings that we applied, which we can then use to
look up the actual video information.