Flesh out some actual popup content
This still isn't a ton of information (and due to various silly mistakes
it took far longer than technically necessary to write), but this gives
us a popup that tells us the title of the video, it's view, like and
dislike count, as well as the first line of the description.

More work can (and should) be done here to brighten this up and add more
things, such as commands, but this is good enough as a proof of
concept.