Include visibility status in the hover popup
As a start to filling up our hover popup with interesting video
information, this includes an indication of whether a video is public,
private or unlisted to the hover popup.

The colors here are still a work in progress (particularly for unlisted,
whose main distinction is not being green or red.

Despite the simplicity of this operation, it still took a bit longer
than I am comfortably willing to admit to realize that hi and h1 are not
in fact the same thing and as a result, CSS rules do not apply as one
would expect.