Update the edit details command
This modifies the command for editing video details in the editor so
that it will accept a provided video_id and look up the details for
that video directly instead of prompting the user first.

This requires overriding a base method slightly in order to do this,
but works well.

A slightly better implementation might be to have the command take a
whole dictionary as an argument so that it could get all the details
from there, since the caller in the hover popup has it.

However, this allows anyone with a video ID to call it, at the expense
of a potentially redundant lookup, which will hit the cache anyway.