Convert the report listener to a normal listener
Our hover event listener did not work when the report was generated to a
panel because a view event listener does not get created for a panel
view.

The solution here is to convert the listener to a regular event
listener, which requires us to do a check to ensure that the hover only
happens in a view with an appropriate setting.