Update the scope that we use to access the API
In order to be able to make modifications, we need to have the user
authorize for a scope that will allow editing. Here we are changing the
scope that we use to ask for things from being read-only to being
read-write (and forcing SSL, one assumes).

In addition to this I also updated the OAuth credentials screen on the
google app dashboard, although it is unclear if that was actually needed
or not.