Fix cache error in the video details code
There was a latent error in the caching code for video details in which
there was a typo in the name of the variable used to delete a video.

This would have caused an error except that we don't (currently) have
any refresh calls anywhere that would have triggered it.