Include a set_video_details request for updating
This implements a new request that can take a `part` and a dictionary of
video information, and get YouTube to update that information. 

Currently this naively uses whatever you give it (and the command in the
next commit makes the data as correct as possible), but this needs to
be refactored. However it's currently past midnight, so that's a job for
tomorrow.

Something to note here is that the YouTube API will delete information
from your video if you say you want it but don't provide a value. Thus
the safest bet is to always provide the required details for any parts
that you ask for, to ensure that you don't accidentally clobber
something.