Include new filtering logic for setting details
As a matter of clarity, it should be the code that actually does an
update and not the code that requests the update that should be
responsible for filtering the input data as appropriate, since that is
very much a low level detail.

This implements a new filtering logic that will filter the values being
used to those that the API allows (based on the documentation).