Fix a silly logging issue
A few days ago the logging code was modified to change the name of the
panel so that we could have a second panel for utility output. At that
point the name of the panel was extended, but in the code responsible
for actually showing the panel when requests, the new name was wrong.

This is now resolved, so that once again when errors occur, we'll know
about it.