Update window settings when updating data
This change is small, but probably important for future sanity. When we
get a result back that indicates that we successfully updated the data,
the result contains the new video data.

This is now used to update the setting that we carry with our original
data so that any future command that works in this window has a proper
record.

In addition, we now also provide some visual feedback that the operation
is commencing.