Fix thumbnail loading
In the world of the exceedingly creepy, the modifications recently made
to allow for editing broke the ability for videos in the editor to
display the correct thumbnail, but only for edited videos.

Ostensibly this is because I was supremely fudging the URL by noticing
that the URL for everything is standardized and based on the ID of the
video, and thus templated it.

However, some quick examinations seem to indicate that the URL for a
video's thumbnails is slightly different depending on whether it was
set when the video was added versus being added in later, even though
it still references the video ID. For example:

https://i.ytimg.com/vi/Trumcw2XraM/sddefault.jpg
https://i9.ytimg.com/vi/Trumcw2XraM/sddefault.jpg?sqp=CITGtf8F&rs=AOn4CLDZfW7ywUFN81bMoYw8BMtteSVPlg

Here the first one is the original thumbnail and the second is the new
one. The real magic appears to be the addition of the query string;
without it even the new URL still displays the previous thumbnail,
despite also being on a different host.

In any case, we now just manually grab the appropriate thumbnail, which
was not previously being done because we weren't using Dotty at the time
and the accessor was rather horrendous.