Store change count and original content in buffers
For the purposes of being able to implement our own version of a dirty
file indicator for the scratch buffers that we use to store our video
details, augment the data that we store when we create a new window.

We store the text that we added to the buffer in one setting and the
current change count in another, so that at a future point we can check
the change count to see if the file has been modified, and if it has
(because undo counts as a change), if the content is the same or not.

This is not a great general purpose solution, but there is a finite
(and small) amount of data that our YouTube views can support, so it
works well here.