Only commit when there are new changes to commit
This implements a routine that checks all of the text views in a details
window to see if there are any changes or not. This is done in
potentially two steps.

First, if the change count on the buffer is the same as it was when the
contents were last updated from a reliable source, then no changes have
been made.

Since making a change and then undoing it counts as two changes (i.e.
the change count never goes backwards), if the change count is different
we check to see if the content looks like it's been modified. 

In the case that the change count is different but the text indicates no
change, we update the stored change count so that the comparison doesn't
need to be done again.

This new routine is used to make the commit command disable itself when
there are no changes to commit.