Make empty edit views block data commits
This small change counts a buffer being empty as being unchanged, so
that it is not possible to commit while the title, body or tags are an
empty string.

This stops potential accidents wherein an undo operation takes a buffer
back to the initial state prior to where the original downloaded data
was inserted.

Generally speaking, only the title absolutely requires a value and the
other fields do not, but in practice there are not really any cases in
which you would not enter at least token description and tags for
things.