Block the key for creating a new empty tab
As an extra convenience since we're in the vicinity of key bindings,
this blocks the new file shortcut key in the same way as we block the
close key so that inside of a YouTube window we can't accidentally
create a new tab via the keyboard.

Since our window has no tabs in it, this is potentially incredibly 
confusing, since it would be nigh on impossible to notice without some
interaction.