Update internal state on commit of new details
This augments the commit command so that if it successfully updates the
data via the API, the internal tracking state is updated to what the new
canonical data is.

This essentially "saves" the current data as the default data so that we
cannot commit again until more changes are made (including an undo).